getIndex returns no double

+++ source/src/utility/Data.h
@@ -54,13 +54,13 @@ public:
 
   void getAllValues(std::vector<double>& all_values, std::vector<size_t>& sampleIDs, size_t varID);
 
-  double getIndex(size_t row, size_t col) const {
+  size_t getIndex(size_t row, size_t col) const {
     if (col < num_cols_no_sparse) {
       return index_data[col * num_rows + row];
     } else {
       // Get data out of sparse storage. -1 because of GenABEL coding.
       size_t idx = (col - num_cols_no_sparse) * num_rows_rounded + row;
-      double result = (((sparse_data[idx / 4] & mask[idx % 4]) >> offset[idx % 4]) - 1);
+      size_t result = (((sparse_data[idx / 4] & mask[idx % 4]) >> offset[idx % 4]) - 1);
       return result;
     }
   }
